[REFACTOR] Seperate methods.
